requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
fastapi-throttle==0.1.0  # rate limiting tanpa Redis
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0